requirements.txt
pygame==2.5.2
PyOpenGL==3.1.7
PyOpenGL==3.1.7
pywin32==306
